Write a consistent allproc copy in a live dump
While creating a live dump, proc and thread objects might be modified
while the allproc list is being written to the dump as well as they
might not reside in memory at the time of the live dump. This results
in kgdb being unable to read such live dump or print information on
threads using `info thread`.

In order to write a consistent state of the allproc list, it must be
guaranteed that no other kernel thread modifies it. Since writing the
live dump might take several minutes, it wouldn't be practical to lock
the allproc list while the live dump is created as it would prevent the
rest of processes from running.

Instead of locking the allproc list for the time of the live dump,
we can lock it only when a physical page containing it is being written
to the live dump, replace the allproc list with a consistent copy of it
and unlock it after the page is written.

The consistent copy of the allproc list should consist of proc and
thread objects that are guaranteed to reside in memory when the live
dump is being written. This can be achieved by allocating wired pages to
copy the proc and thread objects and populate the copied allproc list
with them.

All architectures supported in 14.0-CURRENT that implement livedumps
were changed: amd64, arm, arm64, i386, riscv. Note that powerpc doesn't
implement minidumps (livedumps require minidumps) and mips was removed
in 14.0-CURRENT.