mbuf: Avoid extra mbuf allocations in m_pullup()
For some reason m_pullup() will allocate a new head mbuf and copy the
required bytes there when the buffer belongs to a cluster.  That doesn't
make much sense and it complicates RSS hash calculations for PPPoE, so
modify it to leave the chain alone if we have enough contiguous bytes in
the cluster.