netisr: Add a PPPoE flow type and detect it in ether_demux()
Add a callback which can be used to calculate the RSS hash of a
PPP-encapsulated TCP or UDP flow.  Assuming that the hash is not already
set for an incoming mbuf, this will calculate the RSS hash and dispatch
it to a netisr thread.

Note, IPv6 handling is not implemented yet.

This is just intended as a proof-of-concept.