bus: Try providing a proper topology lock
Giant is implicitly released when the owning thread goes to sleep, which
makes it unsuitable as a general purpose lock.  For instance, if two
threads race to call BUS_RESCAN()->pci_reset_method() on the same bus,
and one of them sleeps while invoking device_detach(), duplicate frees
are possible.

Use a second sleepable lock to cover this case.  For now it is not
recursive, but it may need to be.  This is enough to boot a VM with ACPI
and a PCI bus.